Added support for robot certificates.
Robot certificates defined by the corresponding IGTF profile are recognized when
a chain is being parsed. A new Gridsite certificate type was added to mark the
robot in the parsed structures. When the client presents a robot certificate,
the Apache module exports a new environment variable (GRST_ROBOT_DN) containg
the corresponding subject name of the robot certificate. A simple command that
outputs information about a given chain was added.